# input2/student-management/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Student Management API
  description: API for managing student records
  version: 1.0.0
paths:
  /students:
    get:
      summary: Get all students
      responses:
        '200':
          description: A list of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '201':
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
  /students/{id}:
    get:
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the student to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: A student object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found
    put:
      summary: Update a student by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the student to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found
    delete:
      summary: Delete a student by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the student to delete
          schema:
            type: integer
      responses:
        '204':
          description: Student deleted
        '404':
          description: Student not found
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        grade:
          type: string
      required:
        - name
        - age
        - grade